+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix Yucca Trees not being growable on sand tagged blocks
 - Add Cracked Sand and Sandy Dirt to Sand Block and item tags
 - Make Cypress Saplings use 2 by 2
+- Fix Yucca Leaves Models
 
 # 1.0.5-Beta
 - Workaround for a crash involving fruit leaves decay (No Drops)